Add support for the following:
- Immutable fields: Some resources might required certain properties
to be immutable, meaning that these properties will be used to create
resources but once the resource is created these properties will not be
updatable. If the user tries to updated them, the terraform update
operation will fail before it attempts to update the resource.

- ReadOnly: Some resources have properties which are marked as read only.
This means that the property will not be exposed to the end user in the
tf file and will be marked as computed. Computed properties are properties
that are not used to create a resource but might come back from the API
as extra information. These properties will still be saves in the terraform
state file but won't be considered by terraform when performing diffs.

- Force new: An extension of swagger property is now supported in the api
terraform, allowing properties to be configured as force new. This means
that if a given property flagged as force-new is updated terraform will
delete the previously created resource and create a brand new resource
again (with the new property value)