Add support for default values
- Properties which have attributes containing default values are now
translated properly to terraform when creating the resource schema. There's
a chance that the user might have configured in the swagger file a property
containing both readOnly and default attributes - which is not permitted.
Hence, this use case is handled by the validateFunc that provides details
on how to address the issue. Terraform does not allow default values to be
set on computed attributes, hence the need to cover this use case. Having the
validateFunc can give the user a detailed error description on runtime.
- Better error handling