Create symlink pointing to provider instead of renaming the provider binary
- This will ease the plugin installation process in case multiple
providers benefit from the plugin. Creating a symlink for each provider
will reduce greatly the space needed in terms of memory as just one
binary will be required and all the symlinks will be pointing at the same
plugin. Upon terraform command execution, the only thing needed on
runtime will be an env variable in the for of OTF_VAR_$(PROVIDER_NAME)_SWAGGER_URL
where $(PROVIDER_NAME) is the provider name and the value the url
where the provider swagger file is hosted.