major code refactoring to move away from poc code
- improved quality code by having the OpenAPI Terraform provider rely on
interfaces to set up the provider factory and resource factory. This allows
for different implementations to be used such as OpenAPI v2 and make the
provider evolve seamlessly without any major change in the core code of
the provider by setting up the provider with any specific implementation.
- created openapi v2 implementation compliant with openapi spec interfaces
- provider now uses an OpenAPI Client which is able to send http requests
to the appropriate resource using the resource operation configuration
(e,g: security, parameters, etc)

TODO: need to fix the unit tests