add support for bearer auth scheme
- this allows the plugin to automatically include the bearer into the
header/query instead of expecting that to be passed in from the user
reducing the probability for user errors and more importantly does not
expose internal dependencies to the end user.
- SpecSecurityDefinition has been converted to an interface so multiple
implementation can be provided such as bearer auth and regular security
sechemes applicable to both headers and query auths
- the SpecSecurityDefinition implementation are responsible to building
the value fo the header/query accordingly. For instance, in the case of
the bearer auth in headers the