add support for arrays of different types in the specSchemaDefinitionProperty
- resource v2 now supports different elem types for lists
- the specSchemaDefinitionProperty is an intermediary between swagger and
the openapi provider. this object is used as an abstraction so the openapi
provider does not rely on swagger specific implementations.

- added support for ArrayItemsType field that encapsulates the type
of the items (string, int, float, bool, object) if the specSchemaDefinitionProperty
is of type list
- the SpecSchemaDefinition field will contain either the schema of the
property if the type is object or the type of the items if the type is
list and array item type is object

- created helper functions to be albe to reuse code and unit test better

- if the type is list and the items are primitives the terraform schema
returns is of type schema.schema; otherwise schema.Resource is used.