add support for optional computed properties covering all use cases
optional_computed: # optional property that the default value is NOT known at runtime
  type: "string"
  x-terraform-optional-computed: true
optional_computed_with_default:  # computed property that the default value is known at runtime
  type: "string"
  default: “some known default value”

- Default value is no longer mapped to the terraform schema. The value
is expected to be computed by the API as per the open api spec 2.0 documentation. The
client, in this case Terraform, is not expected to push the default value since
the property will be marked as computed.

The reason is because optional computed properties are now set up as Optional = true
and Computed = true; and terraform does not allow the Default field to be
populated if the property is computed.

- Removed form the validateFunc the check that ensures default is not populated
if the property is computed as that is not allowed by Terraform. Since this
is no longer the case considering the Default field is no longer populated
this validation no longer applies.