update integration tests including newly supported property object_nested_scheme_property
- FYI - integration tests are currently failing due to the payload generated when posting
to the API no including the right formatted json format. For instance, the
following example shows what would be the json format of the object_nested_scheme_property
property. AS you can see, the value is set as an array which should not be,
and instead should just be an object.

"object_nested_scheme_property":[{"name":"","object_property":{"account":"someAccountValue"}}]

VS

"object_nested_scheme_property":{"name":"","object_property":{"account":"someAccountValue"}}

In conclusion, we need to update the logic that builds the payload from
the state file and make sure the nested properties are trasnlated to json
correctly.