update negative case to be positive and document the reason why
- subresouce paths are expecteed to be formed following rubresource restful
best practises, meaning each child should be under a given parent that is
identified with their instance id. However, since service providers may
not necesarely follow this pattern in the paths the provider is not enforcing
it to loosen up how opininated the plugin is. Nonetheless, the resulted subresource
will disregard anything that is in between the parent and the child when
building the parent property id.

See other examples of properly build subresource paths to learn more about
resulted configuration for those.