fix backwards compatible changes introduced after refactoring of getTerraformCompliantPropertyName in previous commits
- Context about the aforementioned change: The change was introduced to be
able to build more user friendly parent property names for subresources that
contain versions. For instance cdn_v1_id instead of cdn_v_1_id

What's in this commit?
- updated getTerraformCompliantPropertyName function to honor preferred name
and not do any translation whatsoever
- properties with a number in the value like string_property_1 used to be
left as is, now they are translated to string_property1. To enable this
release without causing a breaking change the preferred name is no longer
automatically converted to compatible name, letting the service provider
define the the preferred name without doing any automated translation
in the plugin itself. If the preferred name is not terraform compliant name the service
provider will have to update the name in the swagger file to be compliant.