fix todos on TestGetResourceURL_edge_cases
- refactored the test driven to use convey following the projects test
standard
- combined happy and unhappy paths into the same set of tests as there's
no need to have two different blocks of tests. However, documented with a
comment the test that are happy paths and the ones that are not.
- fixed some of the bugs raised where now the internal functions return
an error if the id string contain forwards slashes. The other potential
bugs in todos were deemed not valid bugs and rather working as expected.