add two option on how to handle object types without diffs
- option 1 (object_property_argument): this option does not require any
change in the implementation, just serves as documentation to understand
what actions will be required by the user to get over the diffs after the apply
due to the computed properties in the objects not being handled correctly by
terraform core library. The action require by the user will involve describing
in the configuration the readOnly properties with the values saves in the state
file.
- option 2 (object_property_block): This option will effectively make the objects
be treated the same as how nested structs are processed. Meaning, the internal
representation of the object in the state file will be a list with one elem. To support
this we will need another extension (e,g: x-terraform-object-type) to allow the service provider to
enable this behaviour. As per hashi maintainers this should be avoided unless
is strictly necessary.