remove unnecessary error returned from isPropertyWithNestedObjects
- the logic to check whether the schema is present in object types properties
resides somewhere else. This method is responsible for returning whether
a given property is to be considered with nested properties and if it does not
contain nested properties by definition that should return false not an error.
- updated rest of the code to reflect this refactoring including signature
of shouldUseLegacyTerraformSDKBlockApproachForComplexObjects which after removing the
error from isPropertyWithNestedObjects it did not needed to return an error either.
- updated tests to match new logic and removed no longer needed tests